+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Functions for decryptor passthrough: `DAVESession.canPassthrough`, `DAVESession.setPassthroughMode`
+
+### Fixed:
+
+- Fixed setting an external sender possibly not re-creating the group.
+
 ## [0.1.1] - 2025-03-21
 
 ### Fixed: